added new classes for coffee
added new super class for coffees and the sub classes for different types of coffee

+++ src/assignment3/Fika.java
@@ -1,9 +1,8 @@
 package assignment3;
-import java.util.concurrent.ConcurrentLinkedQueue;
 public class Fika {
 
     public static void main(String[] args) {
-        long duration = 3000;
+        double duration = 3000;
         CoffeeSim sim = new CoffeeSim(duration);
         sim.run();
     }
